requirements.txt
Flask>=0.11
click>=6.6
docker-py>=1.9.0
